moving functions/variables into BaseCharacter
