+++ Content/StarterContent/Maps/TestLevel.umap
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:008dbca392d52f857ff2094328f8a36a45465ecda1df778f2a6f0ab7035e914a
-size 199957
+oid sha256:061afdce0699c79d606bee051c80377b024a7513f38402914f4f87d248bc9263
+size 200261
